Add common file upload
