+++ met-web/src/components/common/FileUpload/Uploader.tsx
@@ -1,10 +1,7 @@
 import React, { useContext } from 'react';
-import { Grid, Stack, Typography, IconButton } from '@mui/material';
+import { Grid, Typography } from '@mui/material';
 import Dropzone, { Accept } from 'react-dropzone';
-import { MetWidgetPaper, WidgetButton } from 'components/common';
 import { FileUploadContext } from './FileUploadContext';
-import LinkIcon from '@mui/icons-material/Link';
-import HighlightOffIcon from '@mui/icons-material/HighlightOff';
 
 interface UploaderProps {
     margin?: number;
@@ -12,73 +9,34 @@ interface UploaderProps {
     height?: string;
     acceptedFormat?: Accept;
 }
-const Uploader = ({
-    margin = 2,
-    helpText = 'Drag and drop some files here, or click to select files',
-    height = '10em',
-    acceptedFormat = { 'application/zip': ['.zip'] },
-}: UploaderProps) => {
-    const { handleAddFile, savedFileName, addedFileName, setAddedFileName } = useContext(FileUploadContext);
+const Uploader = ({ margin = 2, height = '10em', helpText, acceptedFormat }: UploaderProps) => {
+    const { handleAddFile } = useContext(FileUploadContext);
 
-    const existingFile = addedFileName;
-
-    if (existingFile) {
-        return (
-            <>
-                <Dropzone
-                    accept={acceptedFormat}
-                    onDrop={(acceptedFiles) => {
-                        handleAddFile(acceptedFiles);
-                        setAddedFileName(acceptedFiles[0].name);
-                    }}
-                >
-                    {({ getRootProps, getInputProps, open }) => (
-                        <section>
-                            <input {...getInputProps()} />
-                            <WidgetButton onClick={open} sx={{ mb: 1 }}>
-                                Upload Shapefile
-                            </WidgetButton>
-                        </section>
-                    )}
-                </Dropzone>
-                <Grid container direction="row" alignItems="flex-start" justifyContent={'flex-start'} item xs={12}>
-                    <MetWidgetPaper elevation={1} sx={{ width: '100%' }}>
-                        <Grid container direction="row" alignItems={'center'} justifyContent="flex-start">
-                            <Grid item xs>
-                                <Stack spacing={2} direction="row" alignItems="center">
-                                    <LinkIcon color="info" />
-                                    <Typography>{addedFileName ? addedFileName : savedFileName}</Typography>
-                                </Stack>
-                            </Grid>
-                            <IconButton
-                                onClick={() => {
-                                    setAddedFileName('');
-                                    handleAddFile([]);
-                                }}
-                                sx={{ padding: 0, margin: 0 }}
-                                color="inherit"
-                                aria-label="Remove Shapefile"
-                            >
-                                <HighlightOffIcon />
-                            </IconButton>
-                        </Grid>
-                    </MetWidgetPaper>
-                </Grid>
-            </>
-        );
-    }
     return (
         <Dropzone
             accept={acceptedFormat}
             onDrop={(acceptedFiles) => {
                 handleAddFile(acceptedFiles);
-                setAddedFileName(acceptedFiles[0].name);
             }}
         >
-            {({ getRootProps, getInputProps, open }) => (
+            {({ getRootProps, getInputProps }) => (
                 <section>
-                    <input {...getInputProps()} />
-                    <WidgetButton onClick={open}>Upload Shapefile </WidgetButton>
+                    <Grid
+                        {...getRootProps()}
+                        container
+                        direction="row"
+                        alignItems="center"
+                        justifyContent="center"
+                        style={{
+                            border: '1px dashed #606060',
+                            background: '#F2F2F2 0% 0% no-repeat padding-box',
+                            height: height,
+                            cursor: 'pointer',
+                        }}
+                    >
+                        <input {...getInputProps()} multiple={false} accept={'image/*'} />
+                        <Typography m={margin}>{helpText}</Typography>
+                    </Grid>
                 </section>
             )}
         </Dropzone>
